[executor] New HookTask + BasicTask refactor
